Modified to use InspectionTestsBase
Modified test class file name to match contained test class name
