--- RubberduckTests/Inspections/OptionBaseZeroInspectionTests.cs
@@ -1,79 +1,0 @@
-using System.Linq;
-using NUnit.Framework;
-using RubberduckTests.Mocks;
-using System.Threading;
-using Rubberduck.Inspections.Concrete;
-
-namespace RubberduckTests.Inspections
-{
-    [TestFixture]
-    public class OptionBaseZeroInspectionTests
-    {
-        [Test]
-        [Category("Inspections")]
-        public void OptionBaseZeroStatement_ReturnsResult()
-        {
-            const string inputCode =
-                @"Option Base 0";
-
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out _);
-            using (var state = MockParser.CreateAndParse(vbe.Object))
-            {
-
-                var inspection = new RedundantOptionInspection(state);
-                var inspector = InspectionsHelper.GetInspector(inspection);
-                var inspectionResults = inspector.FindIssuesAsync(state, CancellationToken.None).Result;
-
-                Assert.AreEqual(1, inspectionResults.Count());
-            }
-        }
-
-        [Test]
-        [Category("Inspections")]
-        public void OptionBaseOneStatement_NoResult()
-        {
-            var inputCode = "Option Base 1";
-
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out _);
-            using (var state = MockParser.CreateAndParse(vbe.Object))
-            {
-
-                var inspection = new RedundantOptionInspection(state);
-                var inspector = InspectionsHelper.GetInspector(inspection);
-                var inspectionResults = inspector.FindIssuesAsync(state, CancellationToken.None).Result;
-
-                Assert.IsFalse(inspectionResults.Any());
-            }
-        }
-
-        [Test]
-        [Category("Inspections")]
-        public void OptionBaseZeroStatement_Ignored_DoesNotReturnResult()
-        {
-            var inputCode =
-                @"'@Ignore OptionBaseZero
-Option Base 0";
-
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out _);
-            using (var state = MockParser.CreateAndParse(vbe.Object))
-            {
-
-                var inspection = new ObsoleteCallStatementInspection(state);
-                var inspector = InspectionsHelper.GetInspector(inspection);
-                var inspectionResults = inspector.FindIssuesAsync(state, CancellationToken.None).Result;
-
-                Assert.IsFalse(inspectionResults.Any());
-            }
-        }
-
-        [Test]
-        [Category("Inspections")]
-        public void InspectionName()
-        {
-            const string inspectionName = "RedundantOptionInspection";
-            var inspection = new RedundantOptionInspection(null);
-
-            Assert.AreEqual(inspectionName, inspection.Name);
-        }
-    }
-}
